backend/requirements.txt
Flask==2.3.3
Flask-Cors==4.0.0
Pillow==10.0.1
pymongo==4.6.0